backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
moviepy==1.0.3
pydub==0.25.1
boto3==1.34.50
requests==2.31.0
yt-dlp==2024.03.10
python-dotenv==1.0.1
pydantic==2.5.0
